adding liquidate denial of service challenge

+++ test/11-op-reg/TargetFunctions.sol
@@ -8,19 +8,19 @@ import { IHevm, vm } from "@chimera/Hevm.sol";
 abstract contract TargetFunctions is BaseTargetFunctions, Properties {
 
     // gets a random non-zero address from `Setup::addressPool`
-    function _getRandomAddress(uint256 index) internal returns(address addr) {
-        index = between(index, 0, ADDRESS_POOL_LENGTH - 1);
+    function _getRandomAddress(uint8 index) internal returns(address addr) {
+        index = uint8(between(index, 0, ADDRESS_POOL_LENGTH - 1));
         addr = addressPool[index];
     }
 
-    function handler_register(uint256 callerIndex) external {
+    function handler_register(uint8 callerIndex) external {
         address caller = _getRandomAddress(callerIndex);
 
         vm.prank(caller);
         operatorRegistry.register();
     }
 
-    function handler_updateAddress(uint256 callerIndex, uint256 updateIndex) external {
+    function handler_updateAddress(uint8 callerIndex, uint8 updateIndex) external {
         address caller = _getRandomAddress(callerIndex);
         address update = _getRandomAddress(updateIndex);
